# Add allocation optimizations