--- a/src/kemal/route_handler.cr
+++ b/src/kemal/route_handler.cr
@@ -41,7 +41,7 @@ module Kemal
     end
 
     private def radix_path(method, path)
-      "/#{method.downcase}#{path}"
+      '/' + method.downcase + path
     end
 
     private def add_to_radix_tree(method, path, route)