This adds optimizations for variable and string allocations.

This benchmark shows that concatenation is more performant than interpolation to append chars/strings (at least up to 3 variables)
```crystal
require "benchmark"

STR = "string"

Benchmark.ips do |x|
  x.report("1 char with 2 strings concat") do
    '/' + STR + STR
  end

  x.report("2 strings with 1 char concat") do
   STR + STR + '/'
  end

  x.report("3 strings concat") do
   STR + STR + STR
  end

  x.report("2 strings with 1 char interpolation") do
    "/#{STR}#{STR}"
  end

  x.report("2 strings with 1 char concat") do
   STR + STR + '/'                                 
  end

  x.report("1 strings with 1 char interpolation") do
    "/#{STR}"
  end

  x.report("String.build") do
    String.build do |str|
      str << '/'
      str << STR
    end
  end
end
```
Results:
```
       1 char with 2 strings concat  19.26M ( 51.92ns) (± 5.46%)   64 B/op        fastest
       2 strings with 1 char concat  15.21M ( 65.75ns) (± 3.37%)   64 B/op   1.27× slower
                   3 strings concat  15.19M ( 65.82ns) (± 3.81%)   64 B/op   1.27× slower
2 strings with 1 char interpolation   9.23M ( 108.3ns) (± 6.85%)  208 B/op   2.09× slower
       2 strings with 1 char concat  16.73M ( 59.76ns) (± 5.68%)   64 B/op   1.15× slower
1 strings with 1 char interpolation  10.92M ( 91.57ns) (± 4.82%)  208 B/op   1.76× slower
                       String.build  11.32M ( 88.33ns) (± 4.78%)  208 B/op   1.70× slower
```