@straight-shoota 2018-08-23
Most changes look good, but please don't replace string interpolation with concatenation. It might be a little bit more performant in some cases, but it's not worth changing it. There is an issue about this in Crystal-lang/crystal#4964 and eventually it will probably be optimized automatically.