@j8r 2018-08-23
Looking at the benchmark, we can up to double the performance - this is an improvement. If interpolation became more performant in the future, we could change concatenations fairly easily.