@straight-shoota 2018-08-23
Please, let's discuss the string optimizations separately. Most of the other changes are pretty straightforward I think.