@sdogruyol 2018-09-16
I just had time to benchmark this locally. Comparing with `master`  increases performance at around %5 - %10 👍 

**master: 40481.00** (avg) vs **branch: 43524.62 (avg)**