@j8r 2018-09-16
@sdogruyol How have you test it, is this the number of requests?
I've reorganized the changes to two commits.