@j8r 2018-10-30
@sdogruyol anything remains to do on my side?